# Q8_KV: 8-bit quantization type targeting the KV cache