--- a/ggml/src/ggml-quants.c
+++ b/ggml/src/ggml-quants.c
@@ -15216,6 +15216,7 @@ bool ggml_validate_row_data(enum ggml_type type, const void * data, size_t nbyte
         case GGML_TYPE_IQ5_K_R4: break;
         case GGML_TYPE_IQ4_KS_R4: break;
         case GGML_TYPE_Q8_K_R8: break;
+        case GGML_TYPE_Q8_KV:   break;
         case GGML_TYPE_BF16_R16: break;
         case GGML_TYPE_Q4_0_4_4:
         case GGML_TYPE_Q4_0_4_8: