… examples/quantize/quantize.cpp
… ggml/src/iqk/iqk_quantize.h
… include/llama.h
… src/llama.cpp